+++ iti-react/src/__tests__/Inputs/DateInput.test.ts
@@ -1,7 +1,8 @@
-import 'jest'
+import moment from 'moment-timezone'
 import {
     convertJsDateToTimeZone,
     parseJsDateIgnoringTimeZone,
+    dateInputValueFromMoment,
 } from '../../Inputs/DateInput'
 
 test('convertJsDateToTimeZone', () => {
@@ -23,4 +24,25 @@ test('parseJsDateIgnoringTimeZone', () => {
     const m = parseJsDateIgnoringTimeZone(date, 'America/Los_Angeles')
 
     expect(m.toString()).toBe('Wed Nov 13 2019 09:00:00 GMT-0800')
+})
+
+describe('dateInputValueFromMoment', () => {
+    it('does a time zone conversion when includesTime=true', () => {
+        const mo = moment.parseZone('1935-05-26T00:00:00+00:00')
+        const value = dateInputValueFromMoment(mo, {
+            includesTime: true,
+            timeZone: 'America/New_York',
+        })
+
+        expect(value.moment).toBe(mo)
+        expect(value.raw).toBe('5/25/1935 8:00 pm')
+    })
+
+    it('does not do a time zone conversion when includesTime=false', () => {
+        const mo = moment.parseZone('1935-05-26T00:00:00+00:00')
+        const value = dateInputValueFromMoment(mo, { includesTime: false })
+
+        expect(value.moment).toBe(mo)
+        expect(value.raw).toBe('5/26/1935')
+    })
 })